orin-ai/backend/tools/tavily_tools.py
# TODO: Phase 2 — Tavily Search Tool
# Reference: execution_plan.md PROMPT 2.2
# Wraps Tavily for the Researcher agent.

import os
import time
from dotenv import load_dotenv
from tavily import TavilyClient

load_dotenv()

# TODO: initialise TavilyClient with TAVILY_API_KEY


def search_competitors(goal: str) -> dict:
    """
    Extracts product domain from goal string.
    Runs 3 Tavily searches (competitors, pain points, market size).
    Aggregates into {competitors, pain_points, market_context}.
    Exponential backoff retry on TavilyError: 2s → 4s → 8s, max 3 attempts.
    On final failure: returns safe fallback dict so pipeline never hard-fails.
    """
    # TODO: implement
    pass


def search_technology(query: str) -> list[dict]:
    """
    Single Tavily search for technology/library research.
    Used by Developer agent when a library install fails.
    Returns list of {title, url, content}.
    """
    # TODO: implement
    pass


if __name__ == "__main__":
    result = search_competitors("task management app")
    print(result)
