orin-ai/backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
anthropic
openai
groq
tavily-python
e2b-code-interpreter
logfire[fastapi]
sse-starlette
pydantic
python-dotenv
httpx
